migrate-ds: user private group support if UIDs are migrated
UPGs are not created during migration:
   b55c98f1c5b0d46aba3f1792ebd8ecc059173b6a

This commit assumes that if user has uidNumber==gidNumber then it is in
fact an UPG and thus creation of UPG is not prevented.

Doesn't work for cases where uidNumber!=gidNumber but user private group
still exists. In such situation we cannot be sure that it exists unless
a check is done. But this check can be very specific to source LDAP server.

It doesn't cover situation when UIDs are not migrated.

https://fedorahosted.org/freeipa/ticket/2562