WebUI Tests: fixing test_selinuxusermap.py
Fixing test_selinuxusermap.py::test_selinuxusermap::test_mod test,
changing the approach of how to scroll to an item in the page.

This commit just change the workaround of scrolling to the element,
to use a new approach.

As it's explained in https://github.com/mozilla/geckodriver/issues/776,
geckodriver do not recognise that the item is on the screen, so the
an assert fails with the error bellow.

AssertionError: Can't click on checkbox
  label:table[name='memberhost_hostgroup'].table
Message: (39, 972.5) is out of bounds of viewport width (1260) and height (964)